requirements.txt
PyYAML>=6.0
markdown>=3.4.0
Jinja2>=3.1.0 